fixed bug in integer formatting

+++ src/util/fmt.cpp
@@ -32,7 +32,7 @@ namespace fmt {
         n++;
       }
 
-      long y = x;
+      long long y = x;
       do {
         ptr++;
         y /= base;
